# openapi.yaml
openapi: 3.1.0
info:
  title: VaultFS API
  description: High-performance self-hosted object storage with built-in image processing
  version: 0.1.0
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000
    description: Local development
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: "API key (vfs_...)"
  schemas:
...
    ObjectMeta:
      type: object
      properties:
        id:
          type: string
          format: uuid
        bucket:
          type: string
        key:
          type: string
        size:
          type: integer
        content_type:
          type: string
        sha256:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        metadata:
          type: object
...
paths:
...
  /v1/stream/{bucket}/{key}:
    parameters:
      - name: bucket
        in: path
        required: true
        schema:
          type: string
      - name: key
        in: path
        required: true
        schema:
          type: string
    put:
      summary: Stream upload a large file (constant RAM usage)
      tags: [Streaming]
      requestBody:
        required: true
        content:
          "*/*":
            schema:
              type: string
              format: binary
      responses:
        "201":
          description: Object created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ObjectMeta"
    get:
      summary: Stream download a file (constant RAM usage)
      tags: [Streaming]
      responses:
        "200":
          description: Streamed file data
          content:
            "*/*":
              schema:
                type: string
                format: binary
...